01-TODO/todo_site/tasks/urls.py
from rest_framework.routers import DefaultRouter
from django.urls import path, include
from .views import CategoryViewSet, TaskViewSet

router = DefaultRouter()
router.register(r'categories', CategoryViewSet)
router.register(r'tasks', TaskViewSet)

urlpatterns = [
    path('', include(router.urls)),
]
