01-TODO/todo_site/requirements.txt
Django>=5.2
djangorestframework>=3.14
django-apscheduler>=0.6.1
